+++ Sources/RichTextKit/Format/RichTextFormatSidebar.swift
@@ -60,14 +60,13 @@ public struct RichTextFormatSidebar: View {
             SidebarSection(title: nil) {
                 VStack(spacing: 4) {
                     RichTextColorPicker(
-                        icon: .richTextColorForeground,
+                        type: .foreground,
                         value: context.binding(for: .foreground),
                         quickColors: .quickPickerColors)
                     RichTextColorPicker(
-                        icon: .richTextColorBackground,
+                        type: .background,
                         value: context.binding(for: .background),
-                        quickColors: .quickPickerColors,
-                        type: .background)
+                        quickColors: .quickPickerColors)
                 }
             }
             .font(.callout)